rename colorUtils.js
